requirements.txt
mysql-connector-python 8.0.31
pip                    22.0.4
protobuf               3.20.1
setuptools             58.1.0